% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3format.R
\name{d3formatFn}
\alias{d3formatFn}
\title{Return textual javascript source that will return a function that lazily instantiate a formatter
for the given string and use it to format the parameter given to the function.}
\usage{
d3formatFn(formatStr, paramName = "value")
}
\arguments{
\item{formatStr}{the d3.format string format specification}

\item{paramName}{the name used for the value parameter in the resulting format. Defaults to "value"}
}
\description{
Note that the lazy instantiation means that we only create a formatting function once, and then call it repeatedly. This creates a 
small permanent memory footprint containing all the previously requested formatters.
}
\examples{
d3format(".3f") # return a javascript string which will give you a function that formats values in fixed format with 3 decimal points
d3format("e")   # return a javascript string that will give you a function that formats values in engineering format
d3format(".3f", "val") # return a javascript string that will reuse the fixed formatter instantiated in the earlier example, with a param name 'val'

}
\keyword{d3.format}
\keyword{javascript}
