% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{depName}
\alias{depName}
\title{Make up a htmlDependency name based on asset type and name}
\usage{
depName(type, name)
}
\arguments{
\item{type}{Type of asset (valid options are: 'plugins', 'plotter' or 'dataHandler')}

\item{name}{Name of asset}
}
\value{
A htmlDependency name.
}
\description{
Make up a htmlDependency name based on asset type and name
}

