% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{dyArrow}
\alias{dyArrow}
\title{dygraph trade arrow}
\usage{
dyArrow(dygraph, x, text = NULL, tooltip = NULL, direction = c("up",
  "down", "left", "right", "ne", "se", "sw", "nw"), fillColor = "white",
  strokeColor = "black", series = NULL)
}
\arguments{
\item{dygraph}{Dygraph to add arrow to}

\item{x}{Either numeric or date/time for the event, depending on the format of the
x-axis of the dygraph. (For date/time must be a \code{POSIXct} object or another
object convertible to \code{POSIXct} via \code{as.POSIXct})}

\item{direction}{Direction of arrow (up or down)}

\item{fillColor}{Fill color of arrow. This can be of the form "#AABBCC" or
"rgb(255,100,200)" or "yellow". Defaults to white.}

\item{strokeColor}{Stroke color of arrow. This can be of the form "#AABBCC" or
"rgb(255,100,200)" or "yellow". Defaults to black.}

\item{seriesName}{Name of series within data set.}

\item{label}{Label for event. Defaults to blank}
}
\value{
A dygraph with the specified trade arrow.
}
\description{
Add an arrow to a dygraph
}
\examples{
library(dygraphs)
dygraph(presidents, main = "Quarterly Presidential Approval Ratings") \%>\%
  dyArrow("1950-6-30", "Korea", direction = "down", fillColor = "red") \%>\%
  dyArrow("1965-2-09", "Vietnam", direction = "down", fillColor = "red")

dygraph(presidents, main = "Quarterly Presidential Approval Ratings") \%>\%
  dyArrow(c("1950-6-30", "1965-2-09"),
          c("Korea", "Vietnam"),
          direction = "down",
          fillColor = "red")

}

