% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{dyAsset}
\alias{dyAsset}
\title{Add an external JavaScript asset as a dygraph dependency}
\usage{
dyAsset(dygraph, type, name, path, version = "1.0")
}
\arguments{
\item{dygraph}{Dygraph to add asset to}

\item{type}{Type of asset (valid options are: 'plugins', 'plotter' or 'dataHandler')}

\item{name}{Name of asset}

\item{path}{Path to asset JavaScript file}

\item{version}{Asset version (e.g. version of package which provides the asset)}
}
\value{
A dygraph with the specified asset added.
}
\description{
Add an external JavaScript asset as a dygraph dependency
}

