% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlestick.R
\name{dyCandlestick}
\alias{dyCandlestick}
\title{Candlestick plotter for dygraph chart}
\usage{
dyCandlestick(dygraph)
}
\arguments{
\item{dygraph}{Dygraph to draw chart on}
}
\value{
Dygraph with specified candlestick plotter
}
\description{
Draw a candlestick chart.
}
\examples{
library(xts)
data(sample_matrix)
library(dygraphs)
dygraph(sample_matrix) \%>\%
  dyCandlestick()

}

