% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{dyCSS}
\alias{dyCSS}
\title{CSS for dygraph labels and legend}
\usage{
dyCSS(dygraph, css)
}
\arguments{
\item{dygraph}{Dygraph to add CSS styles to}

\item{css}{Path to css file to be used for styling textual elements of the
graph.}
}
\value{
dygraph with additional CSS styles
}
\description{
Apply custom CSS to the text drawn within a dygraph. See the
\href{https://dygraphs.com/css.html}{CSS documentation} on the dygraphs
website for additional details on which styles are available.
}
\note{
See the
  \href{https://rstudio.github.io/dygraphs/gallery-css-styling.html}{CSS
  Styling} article on the dygraphs for R website for additional details.

  Note that CSS styles are global so will affect all dygraphs on a given web
  page. This also implies that for a page with multiple plots you only need
  to specify styles for the first one (alternatively you can just add them
  directly to the page by other means).
}
