% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dyDataAdd}
\alias{dyDataAdd}
\alias{dyDataReplace}
\title{Modify underlying dataset}
\usage{
dyDataAdd(dygraph, name, values)

dyDataReplace(dygraph, name, values)
}
\arguments{
\item{dygraph}{Dygraph object to modify}

\item{name}{Name of series to add or replace}

\item{values}{Vector of values for series}
}
\value{
A new dygraph with the requested data modifications
}
\description{
Add data series to or replace series within the underlying dataset
being plotted. These functions are used in the creation of custom
dySeries functions (e.g. to add smoothing, prediction intervals, or
other transformations).
}

