% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{dyDataHandler}
\alias{dyDataHandler}
\title{Include a dygraph data handler}
\usage{
dyDataHandler(dygraph, name, path, version = "1.0")
}
\arguments{
\item{dygraph}{Dygraph to add data handler to}

\item{name}{Name of data handler}

\item{path}{Path to data handler JavaScript file}

\item{version}{Data handler version (e.g. version of package which provides the
data handler)}
}
\value{
A dygraph with the specified data handler enabled.
}
\description{
Include a dygraph data handler
}
