% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{dyGetPluginOptions}
\alias{dyGetPluginOptions}
\title{Get options of included plugin}
\usage{
dyGetPluginOptions(dygraph, name)
}
\arguments{
\item{dygraph}{Dygraph with plugin to get options of}

\item{name}{Name of plugin}
}
\value{
A list with plugin options.
}
\description{
Get options of included plugin
}

