% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{dygraphs-exports}
\alias{dygraphs-exports}
\alias{\%>\%}
\alias{as.yearmon}
\alias{as.yearqtr}
\title{dygraph exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the dygraphs package to avoid listing the magrittr and
zoo packages as Depends of dygraphs.
}
