% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{dyHasPlugin}
\alias{dyHasPlugin}
\title{Check if plugin was included}
\usage{
dyHasPlugin(dygraph, name)
}
\arguments{
\item{dygraph}{Dygraph to check plugin existance in}

\item{name}{Name of plugin}
}
\value{
TRUE if plugin was included in dygraph, FALSE otherwise.
}
\description{
Check if plugin was included
}

