% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dyInteraction}
\alias{dyInteraction}
\title{dygraph interaction}
\usage{
dyInteraction(dygraph, highlightCircleSize = 3,
  highlightSeriesBackgroundAlpha = 0.5, highlightSeriesOpts = list(),
  showLabelsOnHighlight = TRUE, hideOverlyOnMouseOut = TRUE,
  showRoller = FALSE, rollPeriod = 1)
}
\arguments{
\item{dygraph}{Dygraph to add interactive behaviors to}

\item{highlightCircleSize}{The size in pixels of the dot drawn over
highlighted points.}

\item{highlightSeriesBackgroundAlpha}{Fade the background while highlighting
series. 1=fully visible background (disable fading), 0=hiddden background
(show highlighted series only).}

\item{highlightSeriesOpts}{When set, the options from this list are applied
to the series closest to the mouse pointer for interactive
highlighting. Example: list(strokeWidth = 3). See the documentation on
\code{\link{dySeries}} for additional details on options that can be set.}

\item{showLabelsOnHighlight}{Whether to show the legend upon mouseover.}

\item{hideOverlayOnMouseOut}{Whether to hide the legend when the mouse leaves
the chart area.}

\item{showRoller}{If the rolling average period text box should be shown.}

\item{rollPeriod}{Number of timestamps over which to average data.}
}
\value{
Interaction options
}
\description{
Add interactive behaviors to a dygraph plot.
}

