% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactionModel.R
\name{dyInteractionModel}
\alias{dyInteractionModel}
\title{interactionModel}
\usage{
dyInteractionModel(dygraph)
}
\arguments{
\item{dygraph}{Dygraph to add interactionModel to}
}
\value{
A dygraph that displays a range selector
}
\description{
interactionModel
}
\note{
cyj edit
\href{http://rstudio.github.io/dygraphs/gallery-roll-periods.html}{online
documentation} for additional details and examples.
}

