% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{dyPlotter}
\alias{dyPlotter}
\title{Include a dygraph plotter}
\usage{
dyPlotter(dygraph, name, path, version = "1.0")
}
\arguments{
\item{dygraph}{Dygraph to add plotter to}

\item{name}{Name of plotter}

\item{path}{Path to plotter JavaScript file}

\item{version}{Plotter version (e.g. version of package which provides the plotter)}
}
\value{
A dygraph with the specified plotter enabled.
}
\description{
Include a dygraph plotter
}

