% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{dyPlugin}
\alias{dyPlugin}
\title{Include a dygraph plugin}
\usage{
dyPlugin(dygraph, name, path, options = list(), version = "1.0")
}
\arguments{
\item{dygraph}{Dygraph to add plugin to}

\item{name}{Name of plugin}

\item{path}{Path to plugin JavaScript file}

\item{options}{Named list of options to pass to plugin constructor}

\item{version}{Plugin version (e.g. version of package which provides the plugin)}
}
\value{
A dygraph with the specified plugin enabled.
}
\description{
Include a dygraph plugin
}
\details{
You can use dygraphs plugins to customize the appearance of dygraphs as well
as add new interactive behaviors. For additional information on creating
dygraphs plugins see the
\href{https://github.com/danvk/dygraphs/tree/master/src/plugins}{dygraphs plugins}
documentation.
Once you've created a dygraphs plugins you can use the dyPlugin function to
create an R wrapper for it. See the
\href{https://github.com/rstudio/dygraphs/tree/master/inst/examples/plugins/plugins.R}{example plugins}
for details on how to do this.
}

