% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/series.R
\name{dySeriesData}
\alias{dySeriesData}
\title{Add series data to dygraph}
\usage{
dySeriesData(dygraph, name, values)
}
\arguments{
\item{dygraph}{Dygraph to add data to}

\item{name}{Name of series}

\item{values}{Data values}
}
\value{
Dygraph with additional series data
}
\description{
Add an additional column of series data to a dygraph. This
is typically used in the construction of custom series types
(e.g. log scaled, smoothed, etc.)
}

