% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{dySetPluginOptions}
\alias{dySetPluginOptions}
\title{Set options to included plugin}
\usage{
dySetPluginOptions(dygraph, name, options = list())
}
\arguments{
\item{dygraph}{Dygraph with plugin to set options to}

\item{name}{Name of plugin}

\item{options}{New options to set}
}
\value{
A dygraph with plugin with updated options.
}
\description{
Set options to included plugin
}

