% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{dySliderInput}
\alias{dySliderInput}
\title{dySliderInput}
\usage{
dySliderInput(dygraph, color = "red", strokePattern = c("dashed", "solid",
  "dotted", "dotdash"), animate = FALSE)
}
\arguments{
\item{dygraph}{Dygraph to add plugin to}

\item{color}{Color to draw slider. Defaults to 'red'.}

\item{strokePattern}{Line type for slider. Defaults to 'dashed'. Valid arguments are
'dashed', 'solid', 'dotted', and 'dotdash'.}

\item{animate}{'TRUE' to show simple animation controls with default settings; 
'FALSE' not to; or a custom settings list, such as those created using 'animationOptions'.}
}
\value{
A dygraph with the specified plugin enabled.
}
\description{
dySliderInput
}
\details{
The dySliderInput plugin turns the dyDygraph into a slider input widget. The
user can click on a point along the graph and the graph will place a vertical
line on the graph. The user can also use the animation options to scroll
through points along the graph.
}
\examples{
library(dygraphs)
dygraph(mdeaths) \%>\%
  dySliderInput()

}

