% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dySmoothedSeries}
\alias{dySmoothedSeries}
\title{Smoothed series for dychart}
\usage{
dySmoothedSeries(dychart, name, confidence = TRUE, level = 0.95,
  label = NULL, color = NULL, axis = "y", ...)
}
\arguments{
\item{name}{Name of series within dataset (unamed series can be bound by
using the convention V1, V2, etc.). This can also be a character vector of
length 3 that specifies a set of input series to use as the lower, value,
and upper for a series with a shaded bar drawn around it.}

\item{confidence}{\code{TRUE} to display confidence interval around smooth}

\item{level}{Level of confidence interval to use (defaults to 0.95)}

\item{label}{Label to display for series (uses name if no label defined)}

\item{color}{Color for series. These can be of the form "#AABBCC" or
"rgb(255,100,200)" or "yellow", etc. Note that if you specify a custom
color for one series then you must specify one for all series. If not
specified, equally-spaced points around a color wheel are used.}

\item{axis}{Y-axis to associate the series with ("y" or "y2")}

\item{...}{Additional arguments to pass to \code{\link{dySeries}}}
}
\value{
Dygraph with additional smoothed series
}
\description{
Add a smoothed series to a dychart.
}

