% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{dyUnzoom}
\alias{dyUnzoom}
\title{The dyUnzoom plugin adds an "Unzoom" button to the graph when it's displaying
in a zoomed state (this is a bit more discoverable than the default double-
click gesture for unzooming).}
\usage{
dyUnzoom(dygraph)
}
\arguments{
\item{dygraph}{Dygraph to add plugin to}
}
\value{
Dygraph with Unzoom plugin enabled
}
\description{
The dyUnzoom plugin adds an "Unzoom" button to the graph when it's displaying
in a zoomed state (this is a bit more discoverable than the default double-
click gesture for unzooming).
}
\examples{
library(dygraphs)
dygraph(ldeaths) \%>\%
  dyRangeSelector() \%>\%
  dyUnzoom()

}
