% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dyAnnotation}
\alias{dyAnnotation}
\title{Annotation for dygraph chart}
\usage{
dyAnnotation(x, shortText, text = NULL, series = NULL)
}
\arguments{
\item{x}{Date value indicating where to place the annotation. This should be
of class \code{POSIXct} or convertable to \code{POSIXct}.}

\item{shortText}{Text to overlay on the chart at the location of x}

\item{text}{Additional tooltip text to display on mouse hover}

\item{series}{Name of data series to attach annotation to (not required if
  there is only one series in the chart).}
}
\value{
Annotation definition
}
\description{
Define a text annotation for a data-point on a dygraph chart.
}

