% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dyAnnotation}
\alias{dyAnnotation}
\title{Annotation for dygraph chart}
\usage{
dyAnnotation(dygraph, x, text, tooltip = NULL, series = NULL)
}
\arguments{
\item{x}{Date value indicating where to place the annotation. This should be
of class \code{POSIXct} or convertable to \code{POSIXct}.}

\item{text}{Text to overlay on the chart at the location of x}

\item{tooltip}{Additional tooltip text to display on mouse hover}

\item{series}{Name of data series to attach annotation to}

\item{series}{Series to attach the annotation to. By default, the last
series defined using \code{\link{dySeries}}.}
}
\value{
Dygraph with specified annotation
}
\description{
Define a text annotation for a data-point on a dygraph chart.
}

