% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dyAxis}
\alias{dyAxis}
\title{dygraph axis options}
\usage{
dyAxis(name, label = NULL, ...)
}
\arguments{
\item{name}{Axis name ('x', 'y', or 'y2')}

\item{label}{Label to display for axis (defaults to none)}

\item{...}{Per-axis options to pass directly to dygraphs (see the
  \href{http://dygraphs.com/options.html}{dygraphs documentation} for
  additional details).}
}
\value{
Axis options
}
\description{
Define options for an axis on a dygraph plot.
}

