% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dyAxis}
\alias{dyAxis}
\title{dygraph axis options}
\usage{
dyAxis(dygraph, name, label = NULL, drawGrid = NULL, ...)
}
\arguments{
\item{dygraph}{Plot to add options to}

\item{name}{Name of axis ('x', 'y', or 'y2')}

\item{label}{Label to display for axis (defaults to none)}

\item{drawGrid}{Whether to display gridlines in the chart for this axis (uses
the global default if not specified).}

\item{...}{Additional per-axis options to pass directly to dygraphs (see the
  \href{http://dygraphs.com/options.html}{dygraphs documentation} for
  additional details).}
}
\value{
Interactive dygraph plot
}
\description{
Add per-axis options to a dygraph plot.
}

