% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlestick.R
\name{dyCandlestick}
\alias{dyCandlestick}
\title{Candlestick plotter for dygraph chart}
\usage{
dyCandlestick(dygraph, compress = FALSE)
}
\arguments{
\item{dygraph}{Dygraph to draw chart on}

\item{compress}{If true, compress data yearly, quarterly, monthly, weekly or daily
according to overall amount of bars and/or current zoom level.}
}
\value{
Dygraph with specified candlestick plotter
}
\description{
Draw a candlestick chart.
}
\examples{
library(xts)
data(sample_matrix)
library(dygraphs)
dygraph(sample_matrix) \%>\%
  dyCandlestick()

}

