% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{dyDependency}
\alias{dyDependency}
\title{Add external assets as a dygraph dependency}
\usage{
dyDependency(dygraph, dependency)
}
\arguments{
\item{dygraph}{Dygraph to add dependency to}

\item{dependency}{An HTML dependency}
}
\value{
A dygraph with the specified dependency added.
}
\description{
Add external assets as a dygraph dependency
}
