% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dyEvent}
\alias{dyEvent}
\title{dygraph event line}
\usage{
dyEvent(dygraph, date, label, color = "black", strokePattern = "dashed")
}
\arguments{
\item{dygraph}{Dygraph to add event line to}

\item{date}{Date/time for the event (must be a \code{as.POSIXct} object or
another object convertible to \code{as.POSIXct}). convertable via
\code{as.POSIXct}).}

\item{label}{Label for event.}

\item{color}{Color of event line This can be of the form "#AABBCC" or
  "rgb(255,100,200)" or "yellow". Defaults to black.}
}
\value{
A dygraph with the specified event line.
}
\description{
Add a vertical event line to a dygraph
}
\note{
See the \href{http://jjallaire.github.io/dygraphs/}{online
  documentation} for additional details and examples.
}

