% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dygraph}
\alias{dygraph}
\title{Interactive plot for time series data}
\usage{
dygraph(data, title = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{Time series data (must be an \link[xts]{xts} object or
object which is covertible to \code{xts}).}

\item{title}{Main plot title (optional)}

\item{width}{Width in pixels (optional, defaults to automatic sizing)}

\item{height}{Height in pixels (optional, defaults to automatic sizing)}
}
\value{
Interactive dygraph plot
}
\description{
R interface to interactive time series plotting using the
\href{http://dygraphs.com}{dygraphs} JavaScript library.
}

