% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dygraph}
\alias{dygraph}
\title{Interactive plot for time series data}
\usage{
dygraph(data, title = NULL, series = list(), axes = list(),
  annotations = list(), options = list(), group = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{data}{Time series data (must be an \link[xts]{xts} object or an object
which is covertible to \code{xts}).}

\item{title}{Main plot title (optional)}

\item{series}{Series definition (or list of series definitions) created using
the \code{\link{dySeries}} function. Series can be bound positionally or
explicity using the \code{name} parameter of \code{dySeries}.}

\item{axes}{Axis definition (or list of axis definitions) created using the
the \code{\link{dyAxis}} function.}

\item{annotations}{Annotation (or list of annotations) of specific x values
on the graph created using the \code{\link{dyAnnotation}} function.}

\item{options}{Additional options to pass directly to dygraphs (see the
\href{http://dygraphs.com/options.html}{dygraphs documentation} for
additional details).}

\item{group}{Group to associate this plot with. The x-axis zoom level of
plots within a group is automatically synchronized.}

\item{width}{Width in pixels (optional, defaults to automatic sizing)}

\item{height}{Height in pixels (optional, defaults to automatic sizing)}
}
\value{
Interactive dygraph plot
}
\description{
R interface to interactive time series plotting using the
\href{http://dygraphs.com}{dygraphs} JavaScript library.
}

