% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dygraph}
\alias{dygraph}
\title{dygraph interactive plot for time series data}
\usage{
dygraph(data, main = NULL, xlab = NULL, ylab = NULL, group = NULL,
  width = NULL, height = NULL)
}
\arguments{
\item{data}{Time series data (must be an \link[xts]{xts} object or an object
which is covertible to \code{xts}).}

\item{main}{Main plot title (optional)}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{group}{Group to associate this plot with. The x-axis zoom level of
plots within a group is automatically synchronized.}

\item{width}{Width in pixels (optional, defaults to automatic sizing)}

\item{height}{Height in pixels (optional, defaults to automatic sizing)}
}
\value{
Interactive dygraph plot
}
\description{
R interface to interactive time series plotting using the
\href{http://dygraphs.com}{dygraphs} JavaScript library.
}

