% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dygraphs-exports}
\alias{\%>\%}
\alias{as.yearmon}
\alias{as.yearqtr}
\alias{dygraphs-exports}
\title{dygraph exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the dygraphs package to avoid listing the magrittr and
zoo packages as Depends of dygraphs.
}

