% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dyLegend}
\alias{dyLegend}
\title{dygraph legend}
\usage{
dyLegend(dygraph, show = c("auto", "always", "onmouseover"), width = 250,
  showZeroValues = TRUE, externalDiv = NULL, hideOnMouseOut = TRUE)
}
\arguments{
\item{dygraph}{Dygraph to configure legend options for.}

\item{show}{When to display the legend. Specify "always" to always show the
legend. Specify "onmouseover" to only display it when a user mouses over
the chart. The default behavior is "auto", which results in "always" when
more than one series is plotted and "onmouseover" when only a single series
is plotted.}

\item{width}{Width (in pixels) of the div which shows the legend.}

\item{showZeroValues}{Show zero value labels in the legend.}

\item{externalDiv}{Show data labels in an external div, rather than on the
graph. This value should be a div element id.}

\item{hideOnMouseOut}{Whether to hide the legend when the mouse leaves the
  chart area. This option applies when \code{show} is set to "onmouseover".
  Note that this also affects the hiding of the \code{\link{dyHighlight}}
  on mouse out.}
}
\value{
A dygraph with customized legend options
}
\description{
Configure options for the dygraph series legend.
}

