% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dyOptions}
\alias{dyOptions}
\title{dygraph options}
\usage{
dyOptions(dygraph, drawGrid = TRUE, fillGraph = FALSE, ...)
}
\arguments{
\item{drawGrid}{Whether to display gridlines in the chart. This may also be
set on a \link[=dyAxis]{per-axis basis}.}

\item{fillGraph}{Should the area underneath the graph be filled? This may
also be set on a \link[=dySeries]{per-series basis}}

\item{...}{Additional options to pass directly to dygraphs (see the
  \href{http://dygraphs.com/options.html}{dygraphs documentation} for
  additional details).}
}
\value{
Interactive dygraph plot
}
\description{
Add options to a dygraph plot
}

