% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dyOptions}
\alias{dyOptions}
\title{dygraph options}
\usage{
dyOptions(stackedGraph = FALSE, stepPlot = FALSE, fillGraph = FALSE,
  drawPoints = FALSE, pointSize = 1, ...)
}
\arguments{
\item{stackedGraph}{If set, stack series on top of one another rather than
drawing them independently. The first series specified in the input data
will wind up on top of the chart and the last will be on bottom.}

\item{stepPlot}{When set, display the graph as a step plot instead of a line
plot. This option can also be set on a per-series basis via
\code{\link{dySeries}}.}

\item{fillGraph}{Should the area underneath the graph be filled? This option
is not compatible with error bars. This option can also be set on a
per-series basis via \code{\link{dySeries}}.}

\item{drawPoints}{Draw a small dot at each point, in addition to a line going
through the point. This makes the individual data points easier to see, but
can increase visual clutter in the chart. This option can also be set on a
per-series basis via \code{\link{dySeries}}.}

\item{pointSize}{The size of the dot to draw on each point in pixels. A dot
is always drawn when a point is "isolated", i.e. there is a missing point
on either side of it. This also controls the size of those dots. This
option can also be set on a per-series basis via \code{\link{dySeries}}.}

\item{...}{Additional options to pass directly to dygraphs (see the
  \href{http://dygraphs.com/options.html}{dygraphs documentation} for
  additional details).}
}
\value{
Series options
}
\description{
Add options to a dygraph plot.
}

