% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dyOptions}
\alias{dyOptions}
\title{dygraph options}
\usage{
dyOptions(fillGraph = FALSE, ...)
}
\arguments{
\item{fillGraph}{Should the area underneath the graph be filled? This option
is not compatible with error bars. This option can also be set on a per-series
basis via \code{\link{dySeries}}.}

\item{...}{Additional options to pass directly to dygraphs (see the
  \href{http://dygraphs.com/options.html}{dygraphs documentation} for
  additional details).}
}
\value{
Series options
}
\description{
Add options to a dygraph plot.
}

