% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dyRangeSelector}
\alias{dyRangeSelector}
\title{dygraph interactive range selection and zooming}
\usage{
dyRangeSelector(dygraph, dateWindow = NULL, height = 40,
  fillColor = " #A7B1C4", strokeColor = "#808FAB", keepMouseZoom = TRUE)
}
\arguments{
\item{dygraph}{Dygraph to add range selector to}

\item{dateWindow}{Initially zoom in on a section of the graph. Is a two
element vector [earliest, latest], where earliest/latest objects
convertable via \code{as.POSIXct}.}

\item{height}{Height, in pixels, of the range selector widget. This option
can only be specified at Dygraph creation time.}

\item{fillColor}{The range selector mini plot fill color. This can be of the
form "#AABBCC" or "rgb(255,100,200)" or "yellow". You can also specify ""
to turn off fill.}

\item{strokeColor}{The range selector mini plot stroke color. This can be of
the form "#AABBCC" or "rgb(255,100,200)" or "yellow". You can also specify
"" to turn off stroke.}

\item{keepMouseZoom}{Keep mouse zoom when adding a range selector}
}
\value{
A dygraph that displays a range selector
}
\description{
Add a range selector to the bottom of the chart that allows users to pan and
zoom to various date ranges.
}
\note{
See the
\href{http://rstudio.github.io/dygraphs/gallery-range-selector.html}{online
documentation} for additional details and examples.
}

