% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dySeries}
\alias{dySeries}
\title{dygraph data series options}
\usage{
dySeries(dygraph, name = NULL, label = NULL, fillGraph = NULL,
  strokeWidth = NULL, drawPoints = NULL, pointSize = NULL,
  highlightCircleSize = NULL)
}
\arguments{
\item{dygraph}{Plot to add options to}

\item{name}{Name of series (this can be excluded if there is only one series)}

\item{label}{Label to display for series (default to name)}

\item{fillGraph}{Should the area underneath the graph be filled? (uses the
  global default if not specified).}
}
\value{
Interactive dygraph plot
}
\description{
Add per-series options to a dygraph plot.
}

