% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dySeries}
\alias{dySeries}
\title{dygraph data series options}
\usage{
dySeries(name = NULL, label = name, ...)
}
\arguments{
\item{name}{Name of series within dataset (unamed series can be bound
by order or using the convention V1, V2, etc.).}

\item{label}{Label to display for series (defaults to name)}

\item{...}{Per-series options to pass directly to dygraphs (see the
  \href{http://dygraphs.com/options.html}{dygraphs documentation} for
  additional details).}
}
\value{
Series options
}
\description{
Add per-series options to a dygraph plot.
}

