% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dySeries}
\alias{dySeries}
\title{dygraph data series options}
\usage{
dySeries(name = NULL, label = NULL, color = NULL, axis = "y",
  fillGraph = FALSE, ...)
}
\arguments{
\item{name}{Name of series within dataset (unamed series can be bound by
order or using the convention V1, V2, etc.). This can also be a character
vector of length 3 that specifies a set of input series to use as the
lower, value, and upper values for a series with a shared bar drawn around
it. In this case the \code{label} parameter must also be specified to
provide a label for the aggregate series.}

\item{label}{Label to display for series (uses name if no label defined)}

\item{color}{Color for series. These can be of the form "#AABBCC" or
"rgb(255,100,200)" or "yellow", etc. Note that if you specify a custom
color for one series then you must specify one for all series. If not
specified, equally-spaced points around a color wheel are used.}

\item{axis}{Y-axis to associate the series with ("y" or "y2")}

\item{fillGraph}{Should the area underneath the graph be filled? This option
  is not compatible with error bars.}

\item{...}{Additional options to pass directly to dygraphs (see the
  \href{http://dygraphs.com/options.html}{dygraphs documentation} for
  additional details).}
}
\value{
Series options
}
\description{
Add per-series options to a dygraph plot.
}

