% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dySeries}
\alias{dySeries}
\title{dygraph data series}
\usage{
dySeries(dygraph, name, label = NULL, color = NULL, axis = "y",
  stepPlot = NULL, fillGraph = NULL, drawPoints = NULL,
  pointSize = NULL, strokeWidth = NULL, strokePattern = NULL,
  strokeBorderWidth = NULL, strokeBorderColor = NULL, ...)
}
\arguments{
\item{dygraph}{Dygraph to add a series definition to}

\item{name}{Name of series within dataset (unamed series can be bound by
using the convention V1, V2, etc.). This can also be a character vector of
length 3 that specifies a set of input series to use as the lower, value,
and upper for a series with a shaded bar drawn around it.}

\item{label}{Label to display for series (uses name if no label defined)}

\item{color}{Color for series. These can be of the form "#AABBCC" or
"rgb(255,100,200)" or "yellow", etc. Note that if you specify a custom
color for one series then you must specify one for all series. If not
specified then the global colors option (typically based on equally-spaced
points around a color wheel). Note also that global and per-series color
specification cannot be mixed.}

\item{axis}{Y-axis to associate the series with ("y" or "y2")}

\item{stepPlot}{When set, display the graph as a step plot instead of a line
plot.}

\item{fillGraph}{Should the area underneath the graph be filled? This option
is not compatible with error bars.}

\item{drawPoints}{Draw a small dot at each point, in addition to a line going
through the point. This makes the individual data points easier to see, but
can increase visual clutter in the chart.}

\item{pointSize}{The size of the dot to draw on each point in pixels. A dot
is always drawn when a point is "isolated", i.e. there is a missing point
on either side of it. This also controls the size of those dots.}

\item{strokeWidth}{The width of the lines connecting data points. This can be
used to increase the contrast or some graphs.}

\item{strokePattern}{A custom pattern array where the even index is a draw
and odd is a space in pixels. If null then it draws a solid line. The array
should have a even length as any odd lengthed array could be expressed as a
smaller even length array. This is used to create dashed lines.}

\item{strokeBorderWidth}{Draw a border around graph lines to make crossing
lines more easily distinguishable. Useful for graphs with many lines.}

\item{strokeBorderColor}{Color for the line border used if
  \code{strokeBorderWidth} is set.}

\item{...}{Additional options to pass directly to dygraphs (see the
  \href{http://dygraphs.com/options.html}{dygraphs documentation} for
  additional details).}
}
\value{
Dygraph with additional series
}
\description{
Add a data series to a dygraph plot. Note that options will use the default
global setting (as determined by \code{\link{dyOptions}}) when not specified
explicitly. When no \code{dySeries} is specified for a plot then all series
within the underlying data are plotted.
}
\note{
See the \href{http://jjallaire.github.io/dygraphs/}{online
documentation} for additional details and examples.
}

