% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dySeries}
\alias{dySeries}
\title{dygraph data series options}
\usage{
dySeries(name = NULL, label = name, color = NULL, axis = "y", ...)
}
\arguments{
\item{name}{Name of series within dataset (unamed series can be bound by
order or using the convention V1, V2, etc.).}

\item{label}{Label to display for series (defaults to name)}

\item{color}{Color for series. These can be of the form "#AABBCC" or
"rgb(255,100,200)" or "yellow", etc. Note that if you specify a custom
color for one series then you must specify one for all series. If not
specified, equally-spaced points around a color wheel are used.}

\item{axis}{Y-axis to associate the series with ("y" or "y2")}

\item{...}{Per-series options to pass directly to dygraphs (see the
  \href{http://dygraphs.com/options.html}{dygraphs documentation} for
  additional details).}
}
\value{
Series options
}
\description{
Add per-series options to a dygraph plot.
}

