% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shading.R
\name{dyShading}
\alias{dyShading}
\title{dygraph region shading}
\usage{
dyShading(dygraph, from, to, color = "#EFEFEF")
}
\arguments{
\item{dygraph}{Dygraph to add shading to}

\item{from}{Date/time to shade from (must be a \code{as.POSIXct} object or
another object convertible to \code{as.POSIXct}). convertable via
\code{as.POSIXct}).}

\item{to}{Date/time to shade to (must be a \code{as.POSIXct} object or
another object convertible to \code{as.POSIXct}). convertable via
\code{as.POSIXct}).}

\item{color}{Color of shading. This can be of the form "#AABBCC" or
  "rgb(255,100,200)" or "yellow". Defaults to a very light gray.}
}
\value{
A dygraph with the specified shading
}
\description{
Specify that a region of a dygraph be drawn with a background shading
}
\note{
See the
  \href{http://rstudio.github.io/dygraphs/gallery-shaded-regions.html}{online
  documentation} for additional details and examples.
}

