% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{dyUnzoom}
\alias{dyUnzoom}
\title{dyUnzoom}
\usage{
dyUnzoom(dygraph)
}
\arguments{
\item{dygraph}{Dygraph to add plugin to}
}
\value{
A dygraph with the specified plugin enabled.
}
\description{
dyUnzoom
}
\details{
The dyUnzoom plugin adds an "Unzoom" button to the graph when it's displaying
in a zoomed state (this is a bit more discoverable than the default double-
click gesture for unzooming). Note that this plugin has no options (see 
below for an example with options).
}
\examples{
library(dygraphs)
dygraph(mdeaths) \%>\%
  dyUnzoom()

}

